+++ app/controllers/errors_controller.rb
@@ -1,15 +1,23 @@
 class ErrorsController < ApplicationController
+  
+  # https://guides.rubyonrails.org/v4.2.0/action_controller_overview.html
 
+  layout "errors"
+  
   def not_found
-    render status: 404
+    render status: :not_found
   end
-
-  def unacceptable
-    render status: 422
+ 
+  def unprocessable_entity
+    render status: :unprocessable_entity
   end
 
   def server_error
-    render status: 500
+    render status: :internal_server_error
+  end
+
+  def image_default
+    send_file '/public/img_not_found.jpg', type: 'image/jpg', disposition: 'inline'
   end
 
 end