+++ app/controllers/static_pages_controller.rb
@@ -13,8 +13,4 @@ class StaticPagesController < ApplicationController
   def draw
   end
 
-  def image_default
-    send_file '/public/img_not_found.jpg', type: 'image/jpg', disposition: 'inline'
-  end
-
 end