+++ config/application.rb
@@ -9,15 +9,14 @@ Bundler.require(*Rails.groups)
 module Www
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 7.0
+
+    # catch server errors (404, 500, 422)
     config.exceptions_app = self.routes
-    config.load_defaults 6.0
-	
-	I18n.available_locales = [:en, :es]
-    I18n.default_locale = :en
     config.time_zone = 'Mexico City'
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration can go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded after loading
-    # the framework and any gems in your application.
+    
+    I18n.available_locales = [:en, :es]
+    I18n.default_locale = :es
+    
   end
 end